--- app/[lng]/layout-new.tsx
@@ -1,42 +1,0 @@
-import { dir } from 'i18next';
-import type { Metadata } from 'next';
-import { Geist, Geist_Mono } from 'next/font/google';
-import { languages } from '../../lib/i18n/settings';
-import '../globals.css';
-
-const geistSans = Geist({
-  variable: '--font-geist-sans',
-  subsets: ['latin'],
-});
-
-const geistMono = Geist_Mono({
-  variable: '--font-geist-mono',
-  subsets: ['latin'],
-});
-
-export const metadata: Metadata = {
-  title: 'Next.js i18n Starter',
-  description: 'Next.js i18n with app directory',
-};
-
-export async function generateStaticParams() {
-  return languages.map((lng) => ({ lng }));
-}
-
-export default function RootLayout({
-  children,
-  params: { lng },
-}: Readonly<{
-  children: React.ReactNode;
-  params: { lng: string };
-}>) {
-  return (
-    <html lang={lng} dir={dir(lng)}>
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
-        {children}
-      </body>
-    </html>
-  );
-}
