+++ lib/i18n/index.ts
@@ -1,0 +1,3 @@
+export * from './client';
+export * from './server';
+export * from './settings';
