+++ lib/i18n/server.ts
@@ -10,14 +10,14 @@ const initI18next = async (lng: string, ns: string) => {
     .use(
       resourcesToBackend(
         (language: string, namespace: string) =>
-          import(`../../public/locales/${language}/${namespace}.json`)
+          import(`@/public/locales/${language}/${namespace}.json`)
       )
     )
     .init(getOptions(lng, ns));
   return i18nInstance;
 };
 
-export async function useTranslation(
+export async function useServerTranslation(
   lng: string,
   ns: string,
   options: { keyPrefix?: string } = {}
