+++ lib/index.ts
@@ -1,0 +1,2 @@
+export * from './i18n';
+export * from './utils';
